feat: enhance signup flow with password handling and add login route

+++ routes/auth.js
@@ -13,21 +13,25 @@ const generateOTP = () =>
 
 //request otp 
 router.post("/signup/request-otp", async(req, res) => {
-    const {email} = req.body;
+    const {email, password} = req.body;
 
+    if (!email || !password)
+        return res.status(400).json({ message: "Email and password required" });
 
     const existingUser = await User.findOne({ email });
     if (existingUser)
-    return res.status(400).json({ message: "User already exists" });
+        return res.status(400).json({ message: "User already exists" });
 
     const otp = generateOTP();
     const otpHash = await bcrypt.hash(otp, 10);
+    const passwordHash = await bcrypt.hash(password, 10);
 
     await OTP.findOneAndDelete({email});
 
     await OTP.create({
         email,
         otpHash,
+        passwordHash,
         expiresAt: Date.now() + 5 * 60 * 1000
     });
 
@@ -54,12 +58,47 @@ router.post("/signup/verify-otp", async (req, res) => {
     const valid = await bcrypt.compare(otp, record.otpHash);
     if(!valid) return res.status(400).json({message : "invalid otp"});
 
+    const user = await User.create({
+        email,
+        password: record.passwordHash,
+        isVerified: true
+  });
+
     await OTP.deleteOne({email});
 
-    const token = jwt.sign({email}, process.env.JWT_SECRET, {
-        expiresIn : "1h"
-    });
-    res.json({token})
+    const token = jwt.sign(
+        { userId: user._id },
+        process.env.JWT_SECRET,
+        { expiresIn: "1h" }
+  );
+
+  res.json({token})
 });
+
+
+//post login
+router.post("/login", async(req, res) => {
+    const {email, password} = req.body;
+
+    if (!email || !password)
+        return res.status(400).json({ message: "Email and password required" });
+
+    const user = await User.findOne({email});
+    if(!user)   return res.status(404).json({message:"User not found"});
+
+    const match = await bcrypt.compare(password, user.password);
+    if (!match)
+        return res.status(401).json({ message: "Invalid credentials" });
+
+    const token = jwt.sign(
+        { userId: user._id },
+        process.env.JWT_SECRET,
+        { expiresIn: "1h" }
+    );
+
+    res.json({ token });
+});
+
+
 
 module.exports = router;
